code reorganization and cleanup
