+++ package.json
@@ -37,14 +37,15 @@
     "lodash": "^4.17.20"
   },
   "peerDependencies": {
-    "@jupiterone/integration-sdk-core": "^8.15.0"
+    "@jupiterone/integration-sdk-core": "^8.23.0"
   },
   "devDependencies": {
-    "@jupiterone/integration-sdk-core": "^8.15.0",
-    "@jupiterone/integration-sdk-dev-tools": "^8.15.0",
-    "@jupiterone/integration-sdk-testing": "^8.15.0",
-    "@microsoft/microsoft-graph-types": "^1.28.0",
-    "@microsoft/microsoft-graph-types-beta": "microsoftgraph/msgraph-typescript-typings#beta",
-    "@types/node": "^14"
+    "@jupiterone/integration-sdk-core": "^8.23.0",
+    "@jupiterone/integration-sdk-dev-tools": "^8.23.0",
+    "@jupiterone/integration-sdk-testing": "^8.23.0",
+    "@microsoft/microsoft-graph-types": "^2.24.0",
+    "@microsoft/microsoft-graph-types-beta": "^0.29.0-preview",
+    "@types/node": "^14",
+    "@types/node-fetch": "^2.6.2"
   }
 }
