+++ src/utils/toArray.ts
@@ -1,8 +1,6 @@
-import { isArray } from 'lodash';
-
 /**
  * Convert a string or string array value to a string array
  */
 export function toArray<T>(val: T | T[]): T[] {
-  return isArray(val) ? (val as T[]) : [val as T];
+  return Array.isArray(val) ? (val as T[]) : [val as T];
 }
